requirements.txt
appdirs==1.4.2
beautifulsoup4==4.5.3
bleach==1.5.0
chardet==2.3.0
Django==1.9.12
django-contrib-comments==1.8.0
django-modeltranslation==0.12
filebrowser-safe==0.4.6
future==0.16.0
grappelli-safe==0.4.5
html5lib==0.9999999
Mezzanine==4.1.0
oauthlib==2.0.1
olefile==0.44
packaging==16.8
Pillow==4.0.0
pkg-resources==0.0.0
pyparsing==2.1.10
pytz==2016.10
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
stripe==1.48.1
tzlocal==1.3
